feat: Add empty cart and favorites widgets
The code changes add empty cart and favorites widgets to their respective pages, improving the user experience by displaying a message and an image when the cart or favorites list is empty.

feat: Add lazy loading for images and update product DTO

The code changes include adding lazy loading for images to improve page load performance. Additionally, the product DTO has been updated to include a new typedef for ProductsDto, which takes a list of ProductDto objects and a PaginationDto object.

feat: Update iOS platform version to 13.0

This commit updates the iOS platform version in the Podfile from 12.0 to 13.0. This change ensures compatibility with the latest iOS features and improvements.