feat: Add "unavailableForOrder" message to Russian localization
The code changes include adding the "unavailableForOrder" message to the Russian localization file (strings_ru.i18n.json). This message will be used to display a notification to the user when certain products are unavailable for ordering. This update improves the user experience by providing clear information about the availability of products.